Add client components for comparison and brand search
